+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Next
 ----
 #. Make fields in registration form reorderable.
 #. Set initial values for location and age in registration form, when possible.
+#. Remove hack to django-autopaginate to allow last page as default view. We have our own replacement autopaginate tag now.
 
 1.0
 ---